--- essential-feed-case-studyTests/Feed API/LoadFeedFromRemoteUseCaseTests.swift
@@ -1,124 +1,0 @@
-//
-//  RemoteFeedLoaderTests.swift
-//  essential-feed-case-studyTests
-//
-//  Created by Julius on 30/01/2024.
-//
-
-import XCTest
-import essential_feed_case_study
-
-final class LoadFeedFromRemoteUseCaseTests: XCTestCase {
-    
-    func test_load_deliversErrorOnNon200HTTPURLResponse() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [199, 201, 300, 400, 500]
-        
-        samples.enumerated().forEach { index, code in
-            expect(sut, toCompleteWithResults: failure(.invalidData), when: {
-                let json = makeItemsJSON([])
-                client.complete(withStatusCode: code, data: json, at: index)
-            })
-        }
-    }
-    
-    func test_load_deliversErrorOn200HTTPResponseWithInvalidJSON() {
-        let (sut, client) = makeSUT()
-        
-        expect(sut, toCompleteWithResults: failure(.invalidData), when: {
-            let invalidJSON = Data(_: "invalid json".utf8)
-            client.complete(withStatusCode: 200, data: invalidJSON)
-        })
-    }
-    
-    func test_load_deliversNoItemsOn200HTTPResponseWithEmptyList() {
-        let (sut, client) = makeSUT()
-        
-        expect(sut, toCompleteWithResults: .success([]), when: {
-            let emptyJSONList = makeItemsJSON([])
-            client.complete(withStatusCode: 200, data: emptyJSONList)
-        })
-    }
-    
-    func test_load_deliversItemsOn200HTTPResponseWithJSONItems() {
-        let (sut, client) = makeSUT()
-        let item1 = makeItem(
-            id: UUID(),
-            imageURL: URL(string: "https://a-url.com")!
-        )
-        
-        let item2 = makeItem(
-            id: UUID(),
-            description: "a description",
-            location: "a location",
-            imageURL: URL(string: "https://a-nother-url.com")!
-        )
-        
-        let items = [item1.model, item2.model]
-        
-        expect(sut, toCompleteWithResults: .success(items), when: {
-            let json = makeItemsJSON([item1.json, item2.json])
-            client.complete(withStatusCode: 200, data: json)
-        })
-    }
-    
-    // MARK: -  Helpers
-    
-    private func makeSUT(url: URL = URL(string: "https://a-url.com")!, file: StaticString = #filePath,
-                         line: UInt = #line) -> (sut: RemoteFeedLoader, client: HTTPClientSpy) {
-        let client = HTTPClientSpy()
-        
-        let sut = RemoteFeedLoader(url: url, client: client)
-        trackForMemoryLeaks(sut, file: file, line: line)
-        trackForMemoryLeaks(client, file: file, line: line)
-        return (sut, client)
-    }
-    
-    private func failure(_ error: RemoteFeedLoader.Error) -> RemoteFeedLoader.Result {
-        return .failure(error)
-    }
-    
-    private func makeItem(id: UUID, description: String? = nil, location: String? = nil, imageURL: URL) -> (model: FeedImage, json: [String: Any]) {
-        let item = FeedImage(id: id, description: description, location: location, url: imageURL)
-        let json = [
-            "id": id.uuidString,
-            "description": description,
-            "location": location,
-            "image": imageURL.absoluteString
-        ].compactMapValues { $0 }
-        
-        return (item, json)
-    }
-    
-    private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
-        let json = ["items": items]
-        return try! JSONSerialization.data(withJSONObject: json)
-    }
-    
-    private func expect(_ sut: RemoteFeedLoader, toCompleteWithResults expectedResult: RemoteFeedLoader.Result, when action: () -> Void, file: StaticString = #filePath,
-                        line: UInt = #line) {
-        let exp = expectation(description: "Wait for load completion")
-        
-        sut.load { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedItems), .success(expectedItems)):
-                XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-                
-            case let (.failure(receivedError as RemoteFeedLoader.Error), .failure(expectedError as RemoteFeedLoader.Error)):
-                XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-                
-            default:
-                XCTFail("Expected result \(expectedResult) got \(receivedResult) instead", file: file, line: line)
-            }
-            
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
-        
-    }
-
-}
